Refactored class to handle protocols tree config
