+++ pyworkflow/gui/plugin_manager.py
@@ -27,7 +27,8 @@
 from Tkinter import *
 import webbrowser
 import threading
-from pyworkflow.config import MenuConfig
+
+from pyworkflow.project import MenuConfig
 from pyworkflow.utils.log import ScipionLogger
 from pyworkflow.gui.text import TextFileViewer
 from pyworkflow.gui import *
